requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.11.0
drf-yasg==1.17.1
future==0.18.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
mysqlclient==2.0.1
packaging==20.4
pyparsing==2.4.7
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
whitenoise==5.1.0
zipp==3.1.0
